1.12.0 1. Просмотр, редактирование и удаление звонкового расписания администраторами образовательных организаций 2. Проверка длины звонкового расписания при получении 3. Telegram Web Mini App для редактирования звонкового расписания 4. Обновление firebase-токена даже для неработающей сессии с флагом life=false

+++ tgbot/handlers/school_posts.py
@@ -504,7 +504,7 @@ async def _post_content(message: Message, state: FSMContext):
     """Последовательное написание поста по абзацам"""
 
     async with uow_factory() as uow:
-        school_admin = await check_school_admin(message.from_user.id, uow.school_admin_repository)
+        await check_school_admin(message.from_user.id, uow.school_admin_repository)
 
         if message.text == "Отмена":
             await (await message.answer(".", reply_markup=ReplyKeyboardRemove())).delete()
